removed vertica/common.py and moved all code inside vertica.py
