--- metadata-ingestion/src/datahub/ingestion/source/vertica/common.py
@@ -1,1568 +1,0 @@
-import logging
-import datetime
-import traceback
-from datahub.configuration.common import AllowDenyPattern
-from dataclasses import dataclass
-from datahub.ingestion.api.common import PipelineContext
-from collections import defaultdict
-# from pydantic.fields import Field
-import pydantic
-from pydantic.class_validators import validator
-from sqlalchemy import sql, create_engine
-import json
-from textwrap import dedent
-from typing import (
-    TYPE_CHECKING,
-    Any,
-    Dict,
-    Iterable,
-    List,
-    Optional,
-    Set,
-    Tuple,
-    Type,
-    Union,
-)
-from sqlalchemy.exc import ProgrammingError
-from sqlalchemy.engine.reflection import Inspector
-from datahub.ingestion.api.workunit import MetadataWorkUnit
-from datahub.ingestion.source.sql.sql_common import (
-    BasicSQLAlchemyConfig,
-    SQLAlchemySource,
-    SQLSourceReport,
-    SQLAlchemyConfig,
-    SqlWorkUnit,
-    get_schema_metadata,
-
-)
-# Imports for metadata model classes
-
-from datahub.emitter.mcp import MetadataChangeProposalWrapper
-from datahub.emitter.mce_builder import (
-
-    make_dataset_urn_with_platform_instance,
-    make_tag_urn,
-    dataset_urn_to_key,
-    make_user_urn
-)
-from datahub.emitter.mcp_builder import (
-    wrap_aspect_as_workunit
-)
-
-
-from datahub.metadata.com.linkedin.pegasus2avro.common import StatusClass
-from datahub.metadata.com.linkedin.pegasus2avro.dataset import UpstreamLineage
-from datahub.metadata.com.linkedin.pegasus2avro.metadata.snapshot import DatasetSnapshot
-from datahub.metadata.com.linkedin.pegasus2avro.mxe import MetadataChangeEvent
-from datahub.metadata.schema_classes import (
-    ChangeTypeClass,
-    DataPlatformInstanceClass,
-    DatasetLineageTypeClass,
-    DatasetPropertiesClass,
-    GlobalTagsClass,
-    SubTypesClass,
-    TagAssociationClass,
-    UpstreamClass,
-    ViewPropertiesClass,
-    OwnerClass,
-    OwnershipClass,
-    OwnershipTypeClass,
-)
-from datahub.utilities import config_clean
-
-if TYPE_CHECKING:
-    from datahub.ingestion.source.ge_data_profiler import (
-        GEProfilerRequest,
-    )
-MISSING_COLUMN_INFO = "missing column information"
-logger: logging.Logger = logging.getLogger(__name__)
-
-# Extended SQLSourceReport from sql_common.py to add support for projection , MLModels and Oauth metadata reports .
-
-
-@dataclass
-class VerticaSourceReport(SQLSourceReport):
-    projection_scanned: int = 0
-    models_scanned: int = 0
-    oauth_scanned: int = 0
-
-    def report_entity_scanned(self, name: str, ent_type: str = "table") -> None:
-        """
-        Entity could be a projection or a models or OAuth .
-        """
-
-        if ent_type == "projection":
-            self.projection_scanned += 1
-        elif ent_type == "models":
-            self.models_scanned += 1
-        elif ent_type == "OAuth":
-            self.oauth_scanned += 1
-        else:
-            super().report_entity_scanned(name, ent_type)
-
-# Extended BasicSQLAlchemyConfig to config for projections,models and oauth metadata.
-
-
-class VerticaConfig(BasicSQLAlchemyConfig):
-
-    projection_pattern: AllowDenyPattern = pydantic.Field(
-        default=AllowDenyPattern.allow_all(),
-        description="Regex patterns for projection to filter in ingestion. Specify regex to match the entire projection name in database.schema.projection format. e.g. to match all tables starting with customer in Customer database and public schema, use the regex 'Customer.public.customer.*'",
-    )
-    models_pattern: AllowDenyPattern = pydantic.Field(
-        default=AllowDenyPattern.allow_all(),
-        description="Regex patterns for ml models to filter in ingestion. "
-    )
-    oauth_pattern: AllowDenyPattern = pydantic.Field(
-        default=AllowDenyPattern.allow_all(),
-        description="Regex patterns for OAuth to filter in ingestion. "
-    )
-
-    include_projections: Optional[bool] = pydantic.Field(
-        default=True, description="Whether projections should be ingested."
-    )
-    include_models: Optional[bool] = pydantic.Field(
-        default=True, description="Whether Models should be ingested."
-    )
-    include_oauth: Optional[bool] = pydantic.Field(
-        default=True, description="Whether OAuth should be ingested."
-    )
-    include_view_lineage: Optional[bool] = pydantic.Field(
-        default=True, description="Whether lineages should be ingested for views"
-    )
-    include_projection_lineage: Optional[bool] = pydantic.Field(
-        default=True, description="Whether lineages should be ingested for Projection"
-    )
-
-    # defaults
-    scheme: str = pydantic.Field(default="vertica+vertica_python")
-
-    @validator("host_port")
-    def clean_host_port(cls, v):
-        return config_clean.remove_protocol(v)
-
-
-class VerticaSQLAlchemySource(SQLAlchemySource):
-
-    def __init__(self, config: SQLAlchemyConfig, ctx: PipelineContext, platform: str):
-        # self.platform = platform
-        super(VerticaSQLAlchemySource, self).__init__(config, ctx, platform)
-        self.report: SQLSourceReport = VerticaSourceReport()
-
-    
-
-    def get_workunits(self) -> Iterable[Union[MetadataWorkUnit, SqlWorkUnit]]:
-        sql_config = self.config
-        if logger.isEnabledFor(logging.DEBUG):
-            # If debug logging is enabled, we also want to echo each SQL query issued.
-            sql_config.options.setdefault("echo", True)
-
-        # Extra default SQLAlchemy option for better connection pooling and threading.
-        # https://docs.sqlalchemy.org/en/14/core/pooling.html#sqlalchemy.pool.QueuePool.params.max_overflow
-        if sql_config.profiling.enabled:
-            sql_config.options.setdefault(
-                "max_overflow", sql_config.profiling.max_workers
-            )
-
-        for inspector in self.get_inspectors():
-            profiler = None
-            profile_requests: List["GEProfilerRequest"] = []
-            if sql_config.profiling.enabled:
-                profiler = self.get_profiler_instance(inspector)
-
-            db_name = self.get_db_name(inspector)
-            yield from self.gen_database_containers(
-                inspector=inspector, database=db_name
-            )
-
-            for schema in self.get_allowed_schemas(inspector, db_name):
-                self.add_information_for_schema(inspector, schema)
-
-                yield from self.gen_schema_containers(
-                    inspector=inspector, schema=schema, db_name=db_name
-                )
-
-                if sql_config.include_tables:
-                    yield from self.loop_tables(inspector, schema, sql_config)
-
-                if sql_config.include_views:
-                    yield from self.loop_views(inspector, schema, sql_config)
-
-                if sql_config.include_projections:
-                    yield from self.loop_projections(inspector, schema, sql_config)
-                if sql_config.include_models:
-                    yield from self.loop_models(inspector, schema, sql_config)
-
-                if profiler:
-                    profile_requests += list(
-                        self.loop_profiler_requests(inspector, schema, sql_config)
-                    )
-                    profile_requests += list(
-                        self.loop_profiler_requests_for_projections(inspector, schema, sql_config)
-                    )
-
-            if profiler and profile_requests:
-                yield from self.loop_profiler(
-                    profile_requests, profiler, platform=self.platform
-                )
-
-            oauth_schema = "Entities"
-            if sql_config.include_oauth:
-                yield from self.loop_Oauth(inspector, oauth_schema, sql_config)
-
-        # Clean up stale entities.
-        yield from self.stale_entity_removal_handler.gen_removed_entity_workunits()
-
-    def get_database_properties(
-        self, inspector: Inspector, database: str
-    ) -> Optional[Dict[str, str]]:
-
-        try:
-
-            url = self.config.get_sql_alchemy_url()
-            engine = create_engine(url, **self.config.options)
-
-            cluster_type_qry = sql.text(dedent("""
-                    SELECT 
-                    CASE COUNT(*) 
-                        WHEN 0 
-                        THEN 'Enterprise' 
-                        ELSE 'Eon' 
-                    END AS database_mode 
-                    FROM v_catalog.shards
-                """))
-
-            communal_storage_path = sql.text(dedent("""
-                    SELECT location_path from storage_locations 
-                        WHERE sharing_type = 'COMMUNAL'
-                """))
-
-            cluster_type = ""
-            communal_path = ""
-            cluster_type_res = engine.execute(cluster_type_qry)
-            for each in cluster_type_res:
-                cluster_type = each.database_mode
-                if cluster_type.lower() == 'eon':
-                    for each in engine.execute(communal_storage_path):
-                        communal_path += str(each.location_path) + " | "
-            SUBCLUSTER_SIZE = sql.text(dedent("""
-                            SELECT subclusters.subcluster_name , CAST(sum(disk_space_used_mb // 1024) as varchar(10)) as subclustersize from subclusters  
-                            inner join disk_storage using (node_name) 
-                            group by subclusters.subcluster_name
-                            """))
-            subclusters = " "
-            for data in engine.execute(SUBCLUSTER_SIZE):
-                subclusters += f"{data['subcluster_name']} -- {data['subclustersize']} GB |  "
-
-            cluster__size = sql.text(dedent("""
-                select ROUND(SUM(disk_space_used_mb) //1024 ) as cluster_size
-                from disk_storage
-            """))
-            cluster_size = ""
-            for each in engine.execute(cluster__size):
-                cluster_size = str(each.cluster_size) + " GB"
-
-            return {"cluster_type": cluster_type, "cluster_size": cluster_size, 'subcluster': subclusters,
-                    "communal_storage_path": communal_path}
-
-        except Exception as ex:
-            self.report.report_failure(f"{database}", f"unable to get extra_properties : {ex}")
-
-    def get_schema_properties(
-        self, inspector: Inspector, database: str, schema: str
-    ) -> Optional[Dict[str, str]]:
-
-        try:
-            url = self.config.get_sql_alchemy_url()
-            engine = create_engine(url, **self.config.options)
-            
-            # Projection count
-            projection_count_query = sql.text(dedent("""
-                SELECT 
-                    COUNT(projection_name)  as pc
-                from 
-                    v_catalog.projections 
-                WHERE lower(projection_schema) = '%(schema)s'
-            """ % {"schema": schema}))
-
-            projection_count = None
-            for each in engine.execute(projection_count_query):
-                projection_count = each.pc
-
-            UDL_LANGUAGE = sql.text(dedent("""
-                SELECT lib_name , description 
-                    FROM USER_LIBRARIES
-                WHERE lower(schema_name) = '%(schema)s'
-            """ % {"schema": schema}))
-
-            # UDX list
-            UDX_functions_qry = sql.text(dedent("""
-                SELECT 
-                    function_name 
-                FROM 
-                    USER_FUNCTIONS
-                Where schema_name  = '%(schema)s'
-            """ % {'schema': schema, }))
-            udx_list = ""
-            for each in engine.execute(UDX_functions_qry):
-                udx_list += each.function_name + ", "
-
-            # UDX Language
-            user_defined_library = ""
-            for data in engine.execute(UDL_LANGUAGE):
-                user_defined_library += f"{data['lib_name']} -- {data['description']} |  "
-
-            # print("projection_count: " + str(projection_count)
-            return {"projection_count": str(projection_count),
-                    'udx_list': str(udx_list), 'udx_language': str(user_defined_library)}
-            
-            # return {"projection_count": "projection_count"}
-
-        except Exception as ex:
-            self.report.report_failure(f"{schema}", f"unable to get extra_properties : {ex}")
-            
-
-
-
-
-    def loop_views(
-        self,
-        inspector: Inspector,
-        schema: str,
-        sql_config: SQLAlchemyConfig,
-    ) -> Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]:
-        """
-        loop through all the views in the given schema. This function was re-written due to table_tags variable being added
-        which is getting send to _process_view
-
-        Args:
-            inspector (Inspector): inspector obj of reflaction
-            schema (str): schema name
-            sql_config (SQLAlchemyConfig):
-
-        Returns:
-            Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]:
-
-        Yields:
-            Iterator[Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]]:
-        """
-        try:
-            # table_tags = self.get_extra_tags(inspector, schema, "view")
-            for view in inspector.get_view_names(schema):
-                schema, view = self.standardize_schema_table_names(
-                    schema=schema, entity=view
-                )
-                dataset_name = self.get_identifier(
-                    schema=schema, entity=view, inspector=inspector
-                )
-                dataset_name = self.normalise_dataset_name(dataset_name)
-                self.report.report_entity_scanned(dataset_name, ent_type="view")
-                if not sql_config.view_pattern.allowed(dataset_name):
-                    self.report.report_dropped(dataset_name)
-                    continue
-                try:
-                    yield from self._process_view(
-                        dataset_name=dataset_name,
-                        inspector=inspector,
-                        schema=schema,
-                        view=view,
-                        sql_config=sql_config,
-                       
-                    )
-                except Exception as e:
-                    logger.warning(
-                        f"Unable to ingest view {schema}.{view} due to an exception.\n {traceback.format_exc()}"
-                    )
-                    self.report.report_warning(
-                        f"{schema}.{view}", f"Ingestion error: {e}"
-                    )
-
-                if sql_config.include_view_lineage:
-                    try:
-                        dataset_urn = make_dataset_urn_with_platform_instance(
-                            self.platform,
-                            dataset_name,
-                            self.config.platform_instance,
-                            self.config.env,
-                        )
-
-                        dataset_snapshot = DatasetSnapshot(
-                            urn=dataset_urn,
-                            aspects=[StatusClass(removed=False)],
-                        )
-                        lineage_info = self._get_upstream_lineage_info(dataset_urn, view)
-
-                        if lineage_info is not None:
-                            # Emit the lineage work unit
-                            # upstream_column_props = []
-                            upstream_lineage = lineage_info
-                            lineage_mcpw = MetadataChangeProposalWrapper(
-                                entityType="dataset",
-                                changeType=ChangeTypeClass.UPSERT,
-                                entityUrn=dataset_snapshot.urn,
-                                aspectName="upstreamLineage",
-                                aspect=upstream_lineage,
-                            )
-
-                            lineage_wu = MetadataWorkUnit(
-                                id=f"{self.platform}-{lineage_mcpw.entityUrn}-{lineage_mcpw.aspectName}",
-                                mcp=lineage_mcpw,
-                            )
-                            self.report.report_workunit(lineage_wu)
-                            yield lineage_wu
-
-                    except Exception as e:
-                        logger.warning(
-                            f"Unable to get lineage of view {schema}.{view} due to an exception.\n {traceback.format_exc()}"
-                        )
-                        self.report.report_warning(
-                            f"{schema}.{view}", f"Ingestion error: {e}"
-                        )
-        except Exception as e:
-            self.report.report_failure(f"{schema}", f"Views error: {e}")
-
-    def _process_view(
-        self,
-        dataset_name: str,
-        inspector: Inspector,
-        schema: str,
-        view: str,
-        sql_config: SQLAlchemyConfig,
-      
-    ) -> Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]:
-        """
-        This function is used for performing operation and gets data for every view inside a schema
-
-        Args:
-            dataset_name (str)
-            inspector (Inspector)
-            schema (str): schame name
-            view (str): name of the view to inspect
-            sql_config (SQLAlchemyConfig)
-            table_tags (Dict[str, str], optional) Defaults to dict().
-
-        Returns:
-            Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]
-
-        Yields:
-            Iterator[Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]]
-        """
-        try:
-            columns = inspector.get_columns(view, schema)
-        except KeyError:
-            # For certain types of views, we are unable to fetch the list of columns.
-            self.report.report_warning(
-                dataset_name, "unable to get schema for this view"
-            )
-            schema_metadata = None
-        else:
-            # extra_tags = self.get_extra_tags(inspector, schema, view)
-            extra_tags = dict()
-            schema_fields = self.get_schema_fields(dataset_name, columns, tags=extra_tags)
-            schema_metadata = get_schema_metadata(
-                self.report,
-                dataset_name,
-                self.platform,
-                columns,
-                canonical_schema=schema_fields,
-            )
-        description, properties, _ = self.get_table_properties(inspector, schema, view)
-        try:
-            view_definition = inspector.get_view_definition(view, schema)
-            if view_definition is None:
-                view_definition = ""
-            else:
-                # Some dialects return a TextClause instead of a raw string,
-                # so we need to convert them to a string.
-                view_definition = str(view_definition)
-
-        except NotImplementedError:
-            view_definition = ""
-        properties["view_definition"] = view_definition
-        properties["is_view"] = "True"
-
-        dataset_urn = make_dataset_urn_with_platform_instance(
-            self.platform,
-            dataset_name,
-            self.config.platform_instance,
-            self.config.env,
-        )
-       
-        dataset_snapshot = DatasetSnapshot(
-            urn=dataset_urn,
-            aspects=[StatusClass(removed=False)],
-        )
-        db_name = self.get_db_name(inspector)
-        yield from self.add_table_to_schema_container(dataset_urn, db_name, schema)
-
-
-        # Add view to the checkpoint state
-        self.stale_entity_removal_handler.add_entity_to_state("view", dataset_urn)
-        dataset_properties = DatasetPropertiesClass(
-            name=view,
-            description=description,
-            customProperties=properties,
-        )
-        dataset_snapshot.aspects.append(dataset_properties)
-        if schema_metadata:
-            dataset_snapshot.aspects.append(schema_metadata)
-        mce = MetadataChangeEvent(proposedSnapshot=dataset_snapshot)
-        wu = SqlWorkUnit(id=dataset_name, mce=mce)
-        self.report.report_workunit(wu)
-        yield wu
-        dpi_aspect = self.get_dataplatform_instance_aspect(dataset_urn=dataset_urn)
-        if dpi_aspect:
-            yield dpi_aspect
-        subtypes_aspect = MetadataWorkUnit(
-            id=f"{view}-subtypes",
-            mcp=MetadataChangeProposalWrapper(
-                entityType="dataset",
-                changeType=ChangeTypeClass.UPSERT,
-                entityUrn=dataset_urn,
-                aspectName="subTypes",
-                aspect=SubTypesClass(typeNames=["view"]),
-            ),
-        )
-        self.report.report_workunit(subtypes_aspect)
-        yield subtypes_aspect
-        if "view_definition" in properties:
-            view_definition_string = properties["view_definition"]
-            view_properties_aspect = ViewPropertiesClass(
-                materialized=False, viewLanguage="SQL", viewLogic=view_definition_string
-            )
-            view_properties_wu = MetadataChangeProposalWrapper(
-                entityUrn=dataset_urn,
-                aspect=view_properties_aspect,
-            ).as_workunit()
-            self.report.report_workunit(view_properties_wu)
-            yield view_properties_wu
-        yield from self._get_domain_wu(
-            dataset_name=dataset_name,
-            entity_urn=dataset_urn,
-            entity_type="dataset",
-            sql_config=sql_config,
-        )
-
-    def loop_projections(
-        self,
-        inspector: Inspector,
-        schema: str,
-        sql_config: SQLAlchemyConfig,
-    ) -> Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]:
-        """
-        this function loop through all the projection in the given schema.
-        projection is just like table is introduced by vertica db to store data
-        a super projection is automatically created when a table is created
-
-        Args:
-            inspector (Inspector): inspector obj from reflection
-            schema (str): schema name
-            sql_config (SQLAlchemyConfig): config
-
-        Returns:
-            Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]: [description]
-
-        Yields:
-            Iterator[Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]]: [description]
-        """
-        projections_seen: Set[str] = set()
-
-        try:
-            # table_tags = self.get_extra_tags(inspector, schema, "projection")
-            for projection in inspector.get_projection_names(schema):
-
-                schema, projection = self.standardize_schema_table_names(
-                    schema=schema, entity=projection
-                )
-                dataset_name = self.get_identifier(
-                    schema=schema, entity=projection, inspector=inspector
-                )
-                dataset_name = self.normalise_dataset_name(dataset_name)
-                if dataset_name not in projections_seen:
-                    projections_seen.add(dataset_name)
-                else:
-                    logger.debug(f"{dataset_name} has already been seen, skipping...")
-                    continue
-                self.report.report_entity_scanned(dataset_name, ent_type="projection")
-                if not sql_config.projection_pattern.allowed(dataset_name):
-                    self.report.report_dropped(dataset_name)
-                    continue
-                try:
-                    yield from self._process_projections(
-                        dataset_name, inspector, schema, projection, sql_config
-                    )
-                except Exception as e:
-                    logger.warning(
-                        f"Unable to ingest {schema}.{projection} due to an exception.\n {traceback.format_exc()}"
-                    )
-                    self.report.report_warning(
-                        f"{schema}.{projection}", f"Ingestion error: {e}"
-                    )
-                if sql_config.include_projection_lineage:
-
-                    try:
-                        dataset_urn = make_dataset_urn_with_platform_instance(
-                            self.platform,
-                            dataset_name,
-                            self.config.platform_instance,
-                            self.config.env,
-                        )
-
-                        dataset_snapshot = DatasetSnapshot(
-                            urn=dataset_urn,
-                            aspects=[StatusClass(removed=False)],
-                        )
-                        lineage_info = self._get_upstream_lineage_info_Projection(dataset_urn, projection)
-
-                        if lineage_info is not None:
-                            # Emit the lineage work unit
-
-                            upstream_lineage = lineage_info
-
-                            lineage_mcpw = MetadataChangeProposalWrapper(
-                                entityType="dataset",
-                                changeType=ChangeTypeClass.UPSERT,
-                                entityUrn=dataset_snapshot.urn,
-                                aspectName="upstreamLineage",
-                                aspect=upstream_lineage,
-                            )
-
-                            lineage_wu = MetadataWorkUnit(
-                                id=f"{self.platform}-{lineage_mcpw.entityUrn}-{lineage_mcpw.aspectName}",
-                                mcp=lineage_mcpw,
-                            )
-
-                            self.report.report_workunit(lineage_wu)
-                            yield lineage_wu
-
-                    except Exception as e:
-                        logger.warning(
-                            f"Unable to get lineage of Projection {schema}.{projection} due to an exception.\n {traceback.format_exc()}"
-                        )
-                        self.report.report_warning(
-                            f"{schema}.{projection}", f"Ingestion error: {e}"
-                        )
-        except Exception as e:
-            self.report.report_failure(f"{schema}", f"Projection error: {e}")
-
-    def _process_projections(
-        self,
-        dataset_name: str,
-        inspector: Inspector,
-        schema: str,
-        projection: str,
-        sql_config: SQLAlchemyConfig
-    ) -> Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]:
-        columns = self._get_columns(dataset_name, inspector, schema, projection)
-        dataset_urn = make_dataset_urn_with_platform_instance(
-            self.platform,
-            dataset_name,
-            self.config.platform_instance,
-            self.config.env,
-        )
-        dataset_snapshot = DatasetSnapshot(
-            urn=dataset_urn,
-            aspects=[StatusClass(removed=False)],
-        )
-        # Add table to the checkpoint state
-        self.stale_entity_removal_handler.add_entity_to_state("table", dataset_urn)
-        description, properties, location_urn = self.get_projection_properties(
-            inspector, schema, projection
-        )
-
-        # Tablename might be different from the real table if we ran some normalisation ont it.
-        # Getting normalized table name from the dataset_name
-        # Table is the last item in the dataset name
-        normalised_table = projection
-        splits = dataset_name.split(".")
-        if splits:
-            normalised_table = splits[-1]
-            if properties and normalised_table != projection:
-                properties["original_table_name"] = projection
-
-        dataset_properties = DatasetPropertiesClass(
-            name=normalised_table,
-            description=description,
-            customProperties=properties,
-        )
-        dataset_snapshot.aspects.append(dataset_properties)
-
-        if location_urn:
-            external_upstream_table = UpstreamClass(
-                dataset=location_urn,
-                type=DatasetLineageTypeClass.COPY,
-            )
-            lineage_mcpw = MetadataChangeProposalWrapper(
-                entityType="dataset",
-                changeType=ChangeTypeClass.UPSERT,
-                entityUrn=dataset_snapshot.urn,
-                aspectName="upstreamLineage",
-                aspect=UpstreamLineage(upstreams=[external_upstream_table]),
-            )
-            lineage_wu = MetadataWorkUnit(
-                id=f"{self.platform}-{lineage_mcpw.entityUrn}-{lineage_mcpw.aspectName}",
-                mcp=lineage_mcpw,
-            )
-            self.report.report_workunit(lineage_wu)
-            yield lineage_wu
-
-        extra_tags = self.get_extra_tags(inspector, schema, projection)
-        pk_constraints: dict = inspector.get_pk_constraint(projection, schema)
-        foreign_keys = self._get_foreign_keys(dataset_urn, inspector, schema, projection)
-        schema_fields = self.get_schema_fields(
-            dataset_name, columns, pk_constraints, tags=extra_tags
-        )
-        schema_metadata = get_schema_metadata(
-            self.report,
-            dataset_name,
-            self.platform,
-            columns,
-            pk_constraints,
-            foreign_keys,
-            schema_fields,
-        )
-        dataset_snapshot.aspects.append(schema_metadata)
-        db_name = self.get_db_name(inspector)
-        yield from self.add_table_to_schema_container(dataset_urn, db_name, schema)
-        mce = MetadataChangeEvent(proposedSnapshot=dataset_snapshot)
-        wu = SqlWorkUnit(id=dataset_name, mce=mce)
-        self.report.report_workunit(wu)
-        yield wu
-        dpi_aspect = self.get_dataplatform_instance_aspect(dataset_urn=dataset_urn)
-        if dpi_aspect:
-            yield dpi_aspect
-        subtypes_aspect = MetadataWorkUnit(
-            id=f"{dataset_name}-subtypes",
-            mcp=MetadataChangeProposalWrapper(
-                entityType="dataset",
-                changeType=ChangeTypeClass.UPSERT,
-                entityUrn=dataset_urn,
-                aspectName="subTypes",
-                aspect=SubTypesClass(typeNames=["Projections"]),
-            ),
-        )
-        self.report.report_workunit(subtypes_aspect)
-        yield subtypes_aspect
-
-        yield from self._get_domain_wu(
-            dataset_name=dataset_name,
-            entity_urn=dataset_urn,
-            entity_type="dataset",
-            sql_config=sql_config,
-        )
-
-    # def _get_columns(
-    #     self, dataset_name: str, inspector: Inspector, schema: str, table: str
-    # ) -> List[dict]:
-    #     """
-    #     Return information about columns in `table_name`.
-
-    #     Given a string `table_name` and  `schema`, return
-    #     column information as a list of dicts with these keys:
-
-    #     Args:
-    #         dataset_name (str)
-    #         inspector (Inspector):
-    #         schema (str): schema name
-    #         table (str): table name
-
-    #     Returns:
-    #         List[dict]: list of dictionaries, each representing the definition of
-    #         a database column.
-    #     """
-    #     columns = []
-    #     try:
-    #         columns = inspector.get_columns(table, schema)
-    #         if len(columns) == 0:
-    #             self.report.report_warning(MISSING_COLUMN_INFO, dataset_name)
-    #     except Exception as e:
-    #         self.report.report_warning(
-    #             dataset_name,
-    #             f"unable to get column information due to an error -> {e}",
-    #         )
-    #     return columns
-    def get_projection_properties(
-        self, inspector: Inspector, schema: str, projection: str
-    ) -> Tuple[Optional[str], Dict[str, str], Optional[str]]:
-        """
-        Returns projection related metadata information to show in properties tab
-            eg. projection type like super, segmented etc
-                partition key, segmentation and so on
-
-        Args:
-            inspector (Inspector): inspector obj from reflection engine
-            schema (str): schema name
-            projection (str): projection name
-        Returns:
-            Tuple[Optional[str], Dict[str, str], Optional[str]]: [description]
-        """
-        description: Optional[str] = None
-        properties: Dict[str, str] = {}
-        # The location cannot be fetched generically, but subclasses may override
-        # this method and provide a location.
-        location: Optional[str] = None
-        try:
-            # SQLAlchemy stubs are incomplete and missing this method.
-            # PR: https://github.com/dropbox/sqlalchemy-stubs/pull/223.
-            projection_info: dict = inspector.get_projection_comment(projection, schema)  # type: ignore
-        except NotImplementedError:
-            return description, properties, location
-        except ProgrammingError as pe:
-            logger.debug(
-                f"Encountered ProgrammingError. Retrying with quoted schema name for schema {schema} and table {properties}",
-                pe,
-            )
-            projection_info: dict = inspector.get_projection_comment(properties, f'"{schema}"')  # type: ignore
-        description = projection_info.get("text")
-        if type(description) is tuple:
-            # Handling for value type tuple which is coming for dialect 'db2+ibm_db'
-            description = projection_info["text"][0]
-        # The "properties" field is a non-standard addition to SQLAlchemy's interface.
-        properties = projection_info.get("properties", {})
-        return description, properties, location
-
-    def loop_models(
-        self,
-        inspector: Inspector,
-        schema: str,
-        sql_config: SQLAlchemyConfig,
-    ) -> Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]:
-        """
-        This function is for iterating over the ml models in vertica db
-
-        Args:
-            inspector (Inspector) : inspector obj from reflection engine
-            schema (str): schema name
-            sql_config (SQLAlchemyConfig): config
-
-        Returns:
-            Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]
-
-        Yields:
-            Iterator[Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]]:
-        """
-        models_seen: Set[str] = set()
-        try:
-
-            for models in inspector.get_models_names(schema):
-
-                schema, models = self.standardize_schema_table_names(
-                    schema=schema, entity=models
-                )
-                dataset_name = self.get_identifier(
-                    schema="Entities", entity=models, inspector=inspector
-                )
-
-                dataset_name = self.normalise_dataset_name(dataset_name)
-                if dataset_name not in models_seen:
-                    models_seen.add(dataset_name)
-                else:
-                    logger.debug(f"{dataset_name} has already been seen, skipping...")
-                    continue
-                self.report.report_entity_scanned(dataset_name, ent_type="models")
-                if not sql_config.table_pattern.allowed(dataset_name):
-                    self.report.report_dropped(dataset_name)
-                    continue
-                try:
-                    yield from self._process_models(
-                        dataset_name, inspector, schema, models, sql_config,)
-                except Exception as e:
-                    logger.warning(
-                        f"Unable to ingest {schema}.{models} due to an exception.\n {traceback.format_exc()}"
-                    )
-                    self.report.report_warning(
-                        f"{schema}.{models}", f"Ingestion error: {e}"
-                    )
-        except Exception as e:
-            self.report.report_failure(f"{schema}", f"Model error: {e}")
-
-    def _process_models(
-        self,
-        dataset_name: str,
-        inspector: Inspector,
-        schema: str,
-        table: str,
-        sql_config: SQLAlchemyConfig,
-    ) -> Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]:
-        """
-        To fetch ml models related information of ml_model from vertica db
-        Args:
-            dataset_name (str): dataset name
-            inspector (Inspector): inspector obj from reflection
-            schema (str): schema name entity
-            table (str): name of ml model
-            sql_config (SQLAlchemyConfig)
-
-        Returns:
-            Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]: [description]
-        """
-        columns = []
-        dataset_urn = make_dataset_urn_with_platform_instance(
-            self.platform,
-            dataset_name,
-            self.config.platform_instance,
-            self.config.env,
-        )
-        dataset_snapshot = DatasetSnapshot(
-            urn=dataset_urn,
-            aspects=[StatusClass(removed=False)],
-        )
-        # Add table to the checkpoint state
-        self.stale_entity_removal_handler.add_entity_to_state("model", dataset_urn)
-        description, properties, location_urn = self.get_model_properties(
-            inspector, schema, table
-        )
-        # Tablename might be different from the real table if we ran some normalisation ont it.
-        # Getting normalized table name from the dataset_name
-        # Table is the last item in the dataset name
-        normalised_table = table
-        splits = dataset_name.split(".")
-        if splits:
-            normalised_table = splits[-1]
-            if properties and normalised_table != table:
-                properties["original_table_name"] = table
-        dataset_properties = DatasetPropertiesClass(
-            name=normalised_table,
-            description=description,
-            customProperties=properties,
-        )
-        dataset_snapshot.aspects.append(dataset_properties)
-
-        # extra_tags = self.get_extra_tags(inspector, schema, table)
-      
-
-        schema_fields = self.get_schema_fields(
-            dataset_name, columns
-        )
-
-        schema_metadata = get_schema_metadata(
-            self.report,
-            dataset_name,
-            self.platform,
-            columns,
-            schema_fields,
-        )
-
-        dataset_snapshot.aspects.append(schema_metadata)
-        db_name = self.get_db_name(inspector)
-
-        yield from self.add_table_to_schema_container(dataset_urn, db_name, schema)
-        mce = MetadataChangeEvent(proposedSnapshot=dataset_snapshot)
-        wu = SqlWorkUnit(id=dataset_name, mce=mce)
-        self.report.report_workunit(wu)
-        yield wu
-        dpi_aspect = self.get_dataplatform_instance_aspect(dataset_urn=dataset_urn)
-        if dpi_aspect:
-            yield dpi_aspect
-        subtypes_aspect = MetadataWorkUnit(
-            id=f"{dataset_name}-subtypes",
-            mcp=MetadataChangeProposalWrapper(
-                entityType="dataset",
-                changeType=ChangeTypeClass.UPSERT,
-                entityUrn=dataset_urn,
-                aspectName="subTypes",
-                aspect=SubTypesClass(typeNames=["ML Models"]),
-            ),
-        )
-        self.report.report_workunit(subtypes_aspect)
-        yield subtypes_aspect
-        yield from self._get_domain_wu(
-            dataset_name=dataset_name,
-            entity_urn=dataset_urn,
-            entity_type="dataset",
-            sql_config=sql_config,
-        )
-
-    def get_model_properties(
-        self, inspector: Inspector, schema: str, model: str
-    ) -> Tuple[Optional[str], Dict[str, str], Optional[str]]:
-        """
-        Returns ml models related metadata information to show in properties tab
-            eg. ml model attribute and ml model specification information.
-
-        Args:
-            inspector (Inspector): inspector obj from reflection engine
-            schema (str): schema name
-            model (str): ml model name
-        Returns:
-            Tuple[Optional[str], Dict[str, str], Optional[str]]
-        """
-        description: Optional[str] = None
-        properties: Dict[str, str] = {}
-        # The location cannot be fetched generically, but subclasses may override
-        # this method and provide a location.
-        location: Optional[str] = None
-        try:
-            # SQLAlchemy stubs are incomplete and missing this method.
-            # PR: https://github.com/dropbox/sqlalchemy-stubs/pull/223.
-            table_info: dict = inspector.get_model_comment(model, schema)  # type: ignore
-        except NotImplementedError:
-            return description, properties, location
-        except ProgrammingError as pe:
-            logger.debug(
-                f"Encountered ProgrammingError. Retrying with quoted schema name for schema {schema} and table {model}",
-                pe,
-            )
-            table_info: dict = inspector.get_model_comment(model, f'"{schema}"')  # type: ignore
-        description = table_info.get("text")
-        if type(description) is tuple:
-            # Handling for value type tuple which is coming for dialect 'db2+ibm_db'
-            description = table_info["text"][0]
-        # The "properties" field is a non-standard addition to SQLAlchemy's interface.
-        properties = table_info.get("properties", {})
-        return description, properties, location
-
-    def loop_Oauth(
-        self,
-        inspector: Inspector,
-        schema: str,
-        sql_config: SQLAlchemyConfig,
-    ) -> Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]:
-        """
-        This function is for iterating over the oauth in vertica db
-        Args:
-            inspector (Inspector):
-            schema (str): schema name
-            sql_config (SQLAlchemyConfig): configuration
-
-        Returns:
-            Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]: [description]
-        """
-        oauth_seen: Set[str] = set()
-        try:
-
-            for OAuth in inspector.get_Oauth_names(schema):
-
-                schema, OAuth = self.standardize_schema_table_names(
-                    schema=schema, entity=OAuth
-                )
-                dataset_name = self.get_identifier(
-                    schema=schema, entity=OAuth, inspector=inspector
-                )
-
-                dataset_name = self.normalise_dataset_name(dataset_name)
-                if dataset_name not in oauth_seen:
-                    oauth_seen.add(dataset_name)
-                else:
-                    logger.debug(f"{dataset_name} has already been seen, skipping...")
-                    continue
-                self.report.report_entity_scanned(dataset_name, ent_type="OAuth")
-                if not sql_config.oauth_pattern.allowed(dataset_name):
-                    self.report.report_dropped(dataset_name)
-                    continue
-                try:
-                    yield from self._process_Oauth(
-                        dataset_name, inspector, schema, OAuth, sql_config,)
-                except Exception as e:
-                    logger.warning(
-                        f"Unable to ingest {schema}.{OAuth} due to an exception.\n {traceback.format_exc()}"
-                    )
-                    self.report.report_warning(
-                        f"{schema}.{OAuth}", f"Ingestion error: {e}"
-                    )
-        except Exception as e:
-            self.report.report_failure(f"{schema}", f"Oauth error: {e}")
-
-    def _process_Oauth(
-        self,
-        dataset_name: str,
-        inspector: Inspector,
-        schema: str,
-        OAuth: str,
-        sql_config: SQLAlchemyConfig,
-    ) -> Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]:
-        """
-        To fetch oauth related metadata of oauth from vertica db
-        Args:
-            dataset_name (str): dataset name
-            inspector (Inspector): inspector object from reflection
-            schema (str): schema name
-            OAuth (str): oauth name
-            sql_config (SQLAlchemyConfig): configuration
-        Returns:
-            Iterable[Union[SqlWorkUnit, MetadataWorkUnit]]
-        """
-        columns = []
-        dataset_urn = make_dataset_urn_with_platform_instance(
-            self.platform,
-            dataset_name,
-            self.config.platform_instance,
-            self.config.env,
-        )
-        dataset_snapshot = DatasetSnapshot(
-            urn=dataset_urn,
-            aspects=[StatusClass(removed=False)],
-        )
-        # Add table to the checkpoint state
-        self.stale_entity_removal_handler.add_entity_to_state("OAuth", dataset_urn)
-        description, properties, location_urn = self.get_oauth_properties(
-            inspector, schema, OAuth
-        )
-        # Tablename might be different from the real table if we ran some normalisation ont it.
-        # Getting normalized table name from the dataset_name
-        # Table is the last item in the dataset name
-        normalised_table = OAuth
-        splits = dataset_name.split(".")
-        if splits:
-            normalised_table = splits[-1]
-            if properties and normalised_table != OAuth:
-                properties["original_table_name"] = OAuth
-        dataset_properties = DatasetPropertiesClass(
-            name=normalised_table,
-            description=description,
-            customProperties=properties,
-        )
-        dataset_snapshot.aspects.append(dataset_properties)
-
-        schema_fields = self.get_schema_fields(
-            dataset_name, columns
-        )
-        schema_metadata = get_schema_metadata(
-            self.report,
-            dataset_name,
-            self.platform,
-            columns,
-            schema_fields,
-        )
-        dataset_snapshot.aspects.append(schema_metadata)
-        db_name = self.get_db_name(inspector)
-
-        yield from self.add_table_to_schema_container(dataset_urn, db_name, schema)
-        mce = MetadataChangeEvent(proposedSnapshot=dataset_snapshot)
-        wu = SqlWorkUnit(id=dataset_name, mce=mce)
-        self.report.report_workunit(wu)
-        yield wu
-        dpi_aspect = self.get_dataplatform_instance_aspect(dataset_urn=dataset_urn)
-        if dpi_aspect:
-            yield dpi_aspect
-        subtypes_aspect = MetadataWorkUnit(
-            id=f"{dataset_name}-subtypes",
-            mcp=MetadataChangeProposalWrapper(
-                entityType="dataset",
-                changeType=ChangeTypeClass.UPSERT,
-                entityUrn=dataset_urn,
-                aspectName="subTypes",
-                aspect=SubTypesClass(typeNames=["OAuth"]),
-            ),
-        )
-        self.report.report_workunit(subtypes_aspect)
-        yield subtypes_aspect
-
-        yield from self._get_domain_wu(
-            dataset_name=dataset_name,
-            entity_urn=dataset_urn,
-            entity_type="dataset",
-            sql_config=sql_config,
-        )
-
-    def get_oauth_properties(
-        self, inspector: Inspector, schema: str, model: str
-    ) -> Tuple[Optional[str], Dict[str, str], Optional[str]]:
-        """
-        Returns oauth related metadata information to show in properties tab
-            eg. is_auth_enabled , auth_priority and all auth related info.
-
-        Args:
-            inspector (Inspector): inspector obj from reflection engine
-            schema (str): schema name
-            model (str): ml model name
-        Returns:
-            Tuple[Optional[str], Dict[str, str], Optional[str]]
-        """
-        description: Optional[str] = None
-        properties: Dict[str, str] = {}
-        # The location cannot be fetched generically, but subclasses may override
-        # this method and provide a location.
-        location: Optional[str] = None
-        try:
-            # SQLAlchemy stubs are incomplete and missing this method.
-            # PR: https://github.com/dropbox/sqlalchemy-stubs/pull/223.
-            table_info: dict = inspector.get_oauth_comment(model, schema)  # type: ignore
-        except NotImplementedError:
-            return description, properties, location
-        except ProgrammingError as pe:
-            logger.debug(
-                f"Encountered ProgrammingError. Retrying with quoted schema name for schema {schema} and table {model}",
-                pe,
-            )
-            table_info: dict = inspector.get_oauth_comment(model, f'"{schema}"')  # type: ignore
-        description = table_info.get("text")
-        if type(description) is tuple:
-            # Handling for value type tuple which is coming for dialect 'db2+ibm_db'
-            description = table_info["text"][0]
-        # The "properties" field is a non-standard addition to SQLAlchemy's interface.
-        properties = table_info.get("properties", {})
-        return description, properties, location
-
-    # def get_extra_tags(
-    #     self, inspector: Inspector, schema: str, table: str
-    # ) -> Optional[Dict[str, str]]:
-    #     """
-    #     This function is for projecting the tags at table, view and 
-    #     projection level 
-
-    #     :args: 
-    #         schema: the name of the schema
-    #         table: variable name 
-    #     :return: Dict[str, str]
-    #     """
-    #     try:
-
-    #         tags = inspector._get_extra_tags(table, schema)
-
-    #         return tags
-    #     except Exception as e:
-    #         logger.debug(
-    #             f"{table}: failure in tags extraction for table"
-    #         )
-
-    def gen_tags_aspect_workunit(
-        self, dataset_urn: str, tags_to_add: List[str]
-    ) -> MetadataWorkUnit:
-        """
-        Returns metadata work units object for projecting into tags
-        Args:
-            dataset_urn (str): urn for any dataset
-            tags_to_add (List[str]): tags in form of list of str
-
-        Returns:
-            MetadataWorkUnit
-        """
-        tags = GlobalTagsClass(
-            tags=[TagAssociationClass(tag_to_add) for tag_to_add in tags_to_add]
-        )
-        wu = wrap_aspect_as_workunit("dataset", dataset_urn, "globalTags", tags)
-        self.report.report_workunit(wu)
-        return wu
-
-    def is_dataset_eligible_for_profiling(
-        self,
-        dataset_name: str,
-        sql_config: SQLAlchemyConfig,
-        inspector: Inspector,
-        profile_candidates: Optional[List[str]],
-    ) -> bool:
-        """Return whether table, view or projection is  eligible for profiling 
-
-        Args:
-            dataset_name (str): dataset name
-            sql_config (SQLAlchemyConfig): configuration
-            inspector (Inspector): inspector obj from reflection
-            profile_candidates (Optional[List[str]]): 
-
-        Returns:
-            bool: returns bool
-        """
-        return (
-            sql_config.table_pattern.allowed(dataset_name)
-            and sql_config.profile_pattern.allowed(dataset_name)
-        ) and (
-            sql_config.projection_pattern.allowed(dataset_name)
-            and sql_config.profile_pattern.allowed(dataset_name)
-        ) and (
-            profile_candidates is None
-            or (profile_candidates is not None and dataset_name in profile_candidates)
-        )
-
-    def loop_profiler_requests_for_projections(
-        self,
-        inspector: Inspector,
-        schema: str,
-        sql_config: SQLAlchemyConfig,
-    ) -> Iterable["GEProfilerRequest"]:
-        """Function is used for collecting profiling related information for every projections
-            inside an schema.
-
-        Args: schema: schema name
-
-        Yields:
-            [type]: GEProfilerRequest object
-        """
-        from datahub.ingestion.source.ge_data_profiler import GEProfilerRequest
-        tables_seen: Set[str] = set()
-        profile_candidates = None  # Default value if profile candidates not available.
-        if (
-            sql_config.profiling.profile_if_updated_since_days is not None
-            or sql_config.profiling.profile_table_size_limit is not None
-            or sql_config.profiling.profile_table_row_limit is None
-        ):
-            try:
-                threshold_time: Optional[datetime.datetime] = None
-                if sql_config.profiling.profile_if_updated_since_days is not None:
-                    threshold_time = datetime.datetime.now(
-                        datetime.timezone.utc
-                    ) - datetime.timedelta(
-                        sql_config.profiling.profile_if_updated_since_days
-                    )
-                profile_candidates = self.generate_profile_candidates(
-                    inspector, threshold_time, schema
-                )
-            except NotImplementedError:
-                logger.debug("Source does not support generating profile candidates.")
-        for projection in inspector.get_projection_names(schema):
-
-            schema, projection = self.standardize_schema_table_names(
-                schema=schema, entity=projection
-            )
-            dataset_name = self.get_identifier(
-                schema=schema, entity=projection, inspector=inspector
-            )
-
-            if not self.is_dataset_eligible_for_profiling(
-                dataset_name, sql_config, inspector, profile_candidates
-            ):
-                if self.config.profiling.report_dropped_profiles:
-                    self.report.report_dropped(f"profile of {dataset_name}")
-                continue
-            dataset_name = self.normalise_dataset_name(dataset_name)
-            if dataset_name not in tables_seen:
-                tables_seen.add(dataset_name)
-            else:
-                logger.debug(f"{dataset_name} has already been seen, skipping...")
-                continue
-            missing_column_info_warn = self.report.warnings.get(MISSING_COLUMN_INFO)
-            if (
-                missing_column_info_warn is not None
-                and dataset_name in missing_column_info_warn
-            ):
-                continue
-            (partition, custom_sql) = self.generate_partition_profiler_query(
-                schema, projection, self.config.profiling.partition_datetime
-            )
-            if partition is None and self.is_table_partitioned(
-                database=None, schema=schema, table=projection
-            ):
-                self.report.report_warning(
-                    "profile skipped as partitioned table is empty or partition id was invalid",
-                    dataset_name,
-                )
-                continue
-            if (
-                partition is not None
-                and not self.config.profiling.partition_profiling_enabled
-            ):
-                logger.debug(
-                    f"{dataset_name} and partition {partition} is skipped because profiling.partition_profiling_enabled property is disabled"
-                )
-                continue
-            self.report.report_entity_profiled(dataset_name)
-            logger.debug(
-                f"Preparing profiling request for {schema}, {projection}, {partition}"
-            )
-
-            yield GEProfilerRequest(
-                pretty_name=dataset_name,
-                batch_kwargs=self.prepare_profiler_args(
-                    inspector=inspector,
-                    schema=schema,
-                    table=projection,
-                    partition=partition,
-                    custom_sql=custom_sql,
-                ),
-            )
-
-    def _get_upstream_lineage_info(
-        self, dataset_urn: str, view
-    ) -> Optional[Tuple[UpstreamLineage, Dict[str, str]]]:
-
-        dataset_key = dataset_urn_to_key(dataset_urn)
-        if dataset_key is None:
-            logger.warning(f"Invalid dataset urn {dataset_urn}. Could not get key!")
-            return None
-
-        self._populate_view_lineage(view)
-        dataset_name = dataset_key.name
-        lineage = self.view_lineage_map[dataset_name]
-
-        if not (lineage):
-            logger.debug(f"No lineage found for {dataset_name}")
-            return None
-        upstream_tables: List[UpstreamClass] = []
-
-        for lineage_entry in lineage:
-            # Update the view-lineage
-            upstream_table_name = lineage_entry[0]
-
-            upstream_table = UpstreamClass(
-                dataset=make_dataset_urn_with_platform_instance(
-                    self.platform,
-                    upstream_table_name,
-                    self.config.platform_instance,
-                    self.config.env,
-                ),
-                type=DatasetLineageTypeClass.TRANSFORMED,
-            )
-            upstream_tables.append(upstream_table)
-
-        if upstream_tables:
-
-            logger.debug(
-                f" lineage of '{dataset_name}': {[u.dataset for u in upstream_tables]}"
-            )
-
-            return UpstreamLineage(upstreams=upstream_tables)
-
-        return None
-
-    def _populate_view_lineage(self, view) -> None:
-        """ Collects upstream and downstream lineage information for views .
-
-        Args:
-            view ([type]): name of the view
-
-        """
-
-        url = self.config.get_sql_alchemy_url()
-        logger.debug(f"sql_alchemy_url={url}")
-        engine = create_engine(url, **self.config.options)
-
-        get_refrence_table = sql.text(dedent("""
-            select reference_table_name 
-            from v_catalog.view_tables                                
-            where table_name = '%(view)s'
-        """ % {'view': view}))
-
-        refrence_table = ""
-        for data in engine.execute(get_refrence_table):
-            # refrence_table.append(data)
-            refrence_table = data['reference_table_name']
-
-        view_upstream_lineage_query = sql.text(dedent("""
-            select reference_table_name ,reference_table_schema
-            from v_catalog.view_tables 
-            where table_name = '%(view)s'
-        """ % {'view': view}))
-
-        view_downstream_query = sql.text(dedent("""
-            select table_name ,table_schema
-            from v_catalog.view_tables 
-            where reference_table_name = '%(view)s'
-        """ % {'view': refrence_table}))
-        num_edges: int = 0
-
-        try:
-            self.view_lineage_map = defaultdict(list)
-            for db_row_key in engine.execute(view_downstream_query):
-
-                downstream = f"{db_row_key['table_schema']}.{db_row_key['table_name']}"
-
-                for db_row_value in engine.execute(view_upstream_lineage_query):
-
-                    upstream = f"{db_row_value['reference_table_schema']}.{db_row_value['reference_table_name']}"
-
-                    view_upstream: str = upstream.lower()
-                    view_name: str = downstream.lower()
-                    self.view_lineage_map[view_name].append(
-                        # (<upstream_table_name>, <empty_json_list_of_upstream_table_columns>, <empty_json_list_of_downstream_view_columns>)
-                        (view_upstream, "[]", "[]")
-                    )
-
-                    num_edges += 1
-
-        except Exception as e:
-            self.warn(
-                logger,
-                "view_upstream_lineage",
-                "Extracting the upstream & Downstream view lineage from vertica failed."
-                + f"Please check your permissions. Continuing...\nError was {e}.",
-            )
-
-        logger.info(f"A total of {num_edges} View upstream edges found found for {view}")
-        self.report.num_table_to_view_edges_scanned = num_edges
-
-    def _get_upstream_lineage_info_Projection(
-        self, dataset_urn: str, projection
-    ) -> Optional[Tuple[UpstreamLineage, Dict[str, str]]]:
-
-        dataset_key = dataset_urn_to_key(dataset_urn)
-        if dataset_key is None:
-            logger.warning(f"Invalid dataset urn {dataset_urn}. Could not get key!")
-            return None
-
-        self._populate_projection_lineage(projection)
-        dataset_name = dataset_key.name
-        lineage = self.Projection_lineage_map[dataset_name]
-
-        if not (lineage):
-            logger.debug(f"No lineage found for {dataset_name}")
-            return None
-        upstream_tables: List[UpstreamClass] = []
-
-        for lineage_entry in lineage:
-            # Update the projection-lineage
-            upstream_table_name = lineage_entry[0]
-
-            upstream_table = UpstreamClass(
-                dataset=make_dataset_urn_with_platform_instance(
-                    self.platform,
-                    upstream_table_name,
-                    self.config.platform_instance,
-                    self.config.env,
-                ),
-                type=DatasetLineageTypeClass.TRANSFORMED,
-            )
-            upstream_tables.append(upstream_table)
-
-        if upstream_tables:
-
-            logger.debug(
-                f"lineage of Projection '{dataset_name}': {[u.dataset for u in upstream_tables]}"
-            )
-            return UpstreamLineage(upstreams=upstream_tables)
-        return None
-
-    def _populate_projection_lineage(self, projection) -> None:
-        """
-        Collects upstream and downstream lineage information for views .
-
-        Args:
-            projection ([type]): name of the projection
-
-
-        """
-        url = self.config.get_sql_alchemy_url()
-        logger.debug(f"sql_alchemy_url={url}")
-        engine = create_engine(url, **self.config.options)
-
-        view_upstream_lineage_query = sql.text(dedent("""
-           select basename , schemaname
-           from vs_projections 
-           where name ='%(projection)s'
-        """ % {'projection': projection}))
-
-        num_edges: int = 0
-
-        try:
-            self.Projection_lineage_map = defaultdict(list)
-            for db_row_key in engine.execute(view_upstream_lineage_query):
-                basename = db_row_key['basename']
-                upstream = f"{db_row_key['schemaname']}.{db_row_key['basename']}"
-
-                view_downstream_query = sql.text(dedent("""
-                    select name,schemaname 
-                    from vs_projections 
-                    where basename='%(basename)s'
-                    """ % {'basename': basename}))
-                for db_row_value in engine.execute(view_downstream_query):
-                    downstream = f"{db_row_value['schemaname']}.{db_row_value['name']}"
-                    projection_upstream: str = upstream.lower()
-                    projection_name: str = downstream.lower()
-                    self.Projection_lineage_map[projection_name].append(
-                        # (<upstream_table_name>, <empty_json_list_of_upstream_table_columns>, <empty_json_list_of_downstream_view_columns>)
-                        (projection_upstream, "[]", "[]")
-                    )
-
-                    num_edges += 1
-
-        except Exception as e:
-            self.warn(
-                logger,
-                "Extracting the upstream & downstream Projection lineage from Vertica failed."
-                + f"Please check your permissions. Continuing...\nError was {e}.",
-            )
-
-        logger.info(f"A total of {num_edges} Projection lineage edges found for {projection}.")
-        self.report.num_table_to_view_edges_scanned = num_edges
-    
-    def close(self):
-        self.prepare_for_commit()
